backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
httpx>=0.27
pytest>=8.0
SQLAlchemy>=2.0
alembic>=1.13
asyncpg>=0.29
psycopg2-binary>=2.9
passlib[bcrypt]>=1.7
PyJWT>=2.8
email-validator>=2.1
